+++ scripts/data_miner/transactions.py
@@ -324,7 +324,7 @@ various_txs = [
         ("AI", "0x4fce697c14b1e1310bfa2716fdffa315329ca59c13236ee39bb4f8a04255f557"),    
         ("AI", "0xec60569895b4c6c89cee157e663487681c194e766696cc43103baf6b1ab6191b"),
         # Bylaws
-        ("DAO Bylaws", "0xe38f1714c7310a17c0088d929a10d5e90cf18943e4159d69a078d67e7c977a23"),
+        ("Bylaws", "0xe38f1714c7310a17c0088d929a10d5e90cf18943e4159d69a078d67e7c977a23"),
     #Public Goods
         # Salaries
         ("Salaries", "0xf36d881806444f464a2b648183d8530e3455636cd1294206e307c6829ec8385f"),
@@ -369,15 +369,22 @@ various_txs = [
         ("Eco. Grants", "0x6f5b3d6fd14f54be1973af424024d54758ee3c7ddb0e1a8185b26c02476176fd"),
         ("Gitcoin Grants", "0x9c750f9541a552f731d30439c3fa70ac9a4255b1ea82fb62f604f0d5172d956b"), #Socket
         ("Eco. Grants", "0xab869490b775a12d803aea2c3a171d492aae8002fa72f68e9c306d0dd806e821"),
-        ("Eco. Grants", "0x1d92649577f4b1c6ee087965004c4c0a93893f653c73f5dc3383b7143fca1b6f"),
+        ("Retro Grants", "0x1d92649577f4b1c6ee087965004c4c0a93893f653c73f5dc3383b7143fca1b6f"),
         ("Sponsorship", "0x820af9be1be9c22f945c4b7bc7f4f485780bb2f8cb909aa38c1ffe76e514afeb"),
         ("Eco. Grants", "0xb642434417928afff158d81f92b55f87ca33e71a11a92b207af7a7b0dc2c4ea6"),
+        ("Mini Grants", "0xad7c4c0183d653aa6384883cc54b9b2c9d4d726ae004eb9449f7319b173e6736"),
         ("Mini Grants", "0xc73a2834409b550b1bc052d95860f9088c0dc994187238f86dcac17e6e7f48f4"),
+        ("Hackathons", "0xad7c4c0183d653aa6384883cc54b9b2c9d4d726ae004eb9449f7319b173e6736"),
+        ("Hackathons", "0x4e1885663fe9818d1f59f5edb982ef88d32bad80d72c93f2a2a813a8bc7d6306"),
+        ("Hackathons", "0x4e1885663fe9818d1f59f5edb982ef88d32bad80d72c93f2a2a813a8bc7d6302"),
+        ("Hackathons", "0x4e1885663fe9818d1f59f5edb982ef88d32bad80d72c93f2a2a813a8bc7d6301"),
+        ("$ENS Distibution", "0x6357daddf4fcbbc68d634f092352d2eea48117256f40682f7179d25e2bfa2259"),
     #Public Goods
         ("Grant Platform", "0xdb90045634e3654f169b60ab779da3212a229d0e5597921c0208eb7134f4b5c1"), #Questbook
     #Metagov 
         ("Audit", "0x4a1aed7b7ede77f27116945b2378cfd348244bb73eaa9e1e5129c9e6ca74a081"),
         ("$ENS Vesting", "0x86bb724e316a5c838299959cff8f6c04d77bd9188c6a6faf4ca9e78ca50763bc"),
         ("Security Bounty", "0x53663f861470592a24e05e0bf5c3ade5b26420603c6ac08be46e18d1a37c6e66"),
+        ("DAO Tooling", "0xa8c75013fe5e16c2b3242dfe4d4bb3dfda45ffc3406fcfc05686911f4dd32fba"),
 
 ]